--- a/asv_bench/benchmarks/io/json.py
+++ b/asv_bench/benchmarks/io/json.py
@@ -139,7 +139,6 @@ class ToJSONISO(BaseIO):
 
     def setup(self, orient, frame):
         N = 10 ** 5
-        ncols = 5
         index = date_range("20000101", periods=N, freq="H")
         timedeltas = timedelta_range(start=1, periods=N, freq="s")
         datetimes = date_range(start=1, periods=N, freq="s")