--- a/pandas/_libs/src/ujson/python/objToJSON.c
+++ b/pandas/_libs/src/ujson/python/objToJSON.c
@@ -220,7 +220,7 @@ static TypeContext *createTypeContext(void) {
  *
  * Mutates the provided value directly. Returns 0 on success, non-zero on error.
  */
-static int scaleNanosecToUnit(npy_int64 *value, NPY_DATETIMEUNIT unit) {
+static int64_t scaleNanosecToUnit(npy_int64 *value, NPY_DATETIMEUNIT unit) {
     switch (unit) {
     case NPY_FR_ns:
         break;
@@ -406,36 +406,38 @@ static char *PyUnicodeToUTF8(JSOBJ _obj, JSONTypeContext *tc, size_t *_outLen) {
     return (char *)PyUnicode_AsUTF8AndSize(_obj, (Py_ssize_t *)_outLen);
 }
 
-/* returns a char* and mutates the pointer to *len */
-static char *NpyDateTimeToIso(JSOBJ unused, JSONTypeContext *tc, size_t *len) {
-    npy_datetimestruct dts;
-    int ret_code;
-    int64_t longVal = GET_TC(tc)->longValue;
+/* Converts the int64_t representation of a datetime to ISO; mutates len */
+static char *int64ToIso(int64_t value, NPY_DATETIMEUNIT base, size_t *len) {
+  npy_datetimestruct dts;
+  int ret_code;
+  
+  pandas_datetime_to_datetimestruct(value, NPY_FR_ns, &dts);
+  
+  *len = (size_t)get_datetime_iso_8601_strlen(0, base);
+  char *result = PyObject_Malloc(*len);
 
-    pandas_datetime_to_datetimestruct(longVal, NPY_FR_ns, &dts);
+  if (result == NULL) {
+    PyErr_NoMemory();
+    return NULL;
+  }
 
+  ret_code = make_iso_8601_datetime(&dts, result, *len, base);
+  if (ret_code != 0) {
+    PyErr_SetString(PyExc_ValueError,
+		    "Could not convert datetime value to string");
+    PyObject_Free(result);
+  }
+
+  // Note that get_datetime_iso_8601_strlen just gives a generic size
+  // for ISO string conversion, not the actual size used
+  *len = strlen(result);
+  return result;
+}
+
+/* JSON callback. returns a char* and mutates the pointer to *len */
+static char *NpyDateTimeToIsoCallback(JSOBJ unused, JSONTypeContext *tc, size_t *len) {
     NPY_DATETIMEUNIT base = ((PyObjectEncoder *)tc->encoder)->datetimeUnit;
-    *len = (size_t)get_datetime_iso_8601_strlen(0, base);
-    char *result = PyObject_Malloc(*len);
-
-    if (result == NULL) {
-        PyErr_NoMemory();
-        ((JSONObjectEncoder *)tc->encoder)->errorMsg = "";
-        return NULL;
-    }
-
-    ret_code = make_iso_8601_datetime(&dts, result, *len, base);
-    if (ret_code != 0) {
-        PyErr_SetString(PyExc_ValueError,
-                        "Could not convert datetime value to string");
-        ((JSONObjectEncoder *)tc->encoder)->errorMsg = "";
-        PyObject_Free(result);
-    }
-
-    // Note that get_datetime_iso_8601_strlen just gives a generic size
-    // for ISO string conversion, not the actual size used
-    *len = strlen(result);
-    return result;
+    return int64ToIso(GET_TC(tc)->longValue, base, len);
 }
 
 static npy_datetime NpyDateTimeToEpoch(npy_datetime dt, NPY_DATETIMEUNIT base) {
@@ -443,25 +445,20 @@ static npy_datetime NpyDateTimeToEpoch(npy_datetime dt, NPY_DATETIMEUNIT base) {
     return dt;
 }
 
-static char *PyDateTimeToIso(JSOBJ obj, JSONTypeContext *tc, size_t *len) {
+/* Convert PyDatetime To ISO C-string. mutates len */
+static char *PyDateTimeToIso(PyObject *obj, NPY_DATETIMEUNIT base, size_t *len) {
     npy_datetimestruct dts;
     int ret;
-
-    if (!PyDateTime_Check(obj)) {
-        // TODO: raise TypeError
-    }
-
+    
     ret = convert_pydatetime_to_datetimestruct(obj, &dts);
     if (ret != 0) {
         if (!PyErr_Occurred()) {
             PyErr_SetString(PyExc_ValueError,
                             "Could not convert PyDateTime to numpy datetime");
         }
-        ((JSONObjectEncoder *)tc->encoder)->errorMsg = "";
         return NULL;
     }
-
-    NPY_DATETIMEUNIT base = ((PyObjectEncoder *)tc->encoder)->datetimeUnit;
+    
     *len = (size_t)get_datetime_iso_8601_strlen(0, base);
     char *result = PyObject_Malloc(*len);
     ret = make_iso_8601_datetime(&dts, result, *len, base);
@@ -470,7 +467,6 @@ static char *PyDateTimeToIso(JSOBJ obj, JSONTypeContext *tc, size_t *len) {
         PRINTMARK();
         PyErr_SetString(PyExc_ValueError,
                         "Could not convert datetime value to string");
-        ((JSONObjectEncoder *)tc->encoder)->errorMsg = "";
         PyObject_Free(result);
         return NULL;
     }
@@ -479,6 +475,17 @@ static char *PyDateTimeToIso(JSOBJ obj, JSONTypeContext *tc, size_t *len) {
     // for ISO string conversion, not the actual size used
     *len = strlen(result);
     return result;
+}
+
+/* JSON callback */
+static char *PyDateTimeToIsoCallback(JSOBJ obj, JSONTypeContext *tc, size_t *len) {
+
+    if (!PyDateTime_Check(obj)) {
+        // TODO: raise TypeError
+    }
+
+    NPY_DATETIMEUNIT base = ((PyObjectEncoder *)tc->encoder)->datetimeUnit;
+    return PyDateTimeToIso(obj, base, len);
 }
 
 static npy_datetime PyDateTimeToEpoch(PyObject *obj, NPY_DATETIMEUNIT base) {
@@ -1563,57 +1570,35 @@ char **NpyArr_encodeLabels(PyArrayObject *labels, PyObjectEncoder *enc,
             cLabel = (char *)PyUnicode_AsUTF8(iso);
             Py_DECREF(iso);
             len = strlen(cLabel);
-        } else if (PyTypeNum_ISDATETIME(type_num) || PyDateTime_Check(item) ||
-                   PyDate_Check(item)) {
-            PyObject *ts = PyObject_CallFunction(cls_timestamp, "(O)", item);
-            if (ts == NULL) {
-                Py_DECREF(item);
-                NpyArr_freeLabels(ret, num);
-                ret = 0;
-                break;
-            }
+        } else if (PyTypeNum_ISDATETIME(type_num)) {
+	  NPY_DATETIMEUNIT base = enc->datetimeUnit;
+	  int64_t longVal;
+	  PyArray_VectorUnaryFunc *castfunc =
+            PyArray_GetCastFunc(PyArray_DescrFromType(type_num), NPY_INT64);
+	  if (!castfunc) {
+            PyErr_Format(PyExc_ValueError, "Cannot cast numpy dtype %d to long",
+                         enc->npyType);
+	  }
+	  castfunc(dataptr, &longVal, 1, NULL, NULL);
+	  if (enc->datetimeIso) {
+	    cLabel = int64ToIso(longVal, base, &len);
+	  } else {
+	    scaleNanosecToUnit(&longVal, base);
 
-            if (enc->datetimeIso) {
-                PyObject *iso = PyObject_CallMethod(ts, "isoformat", NULL);
-                Py_DECREF(ts);
-                if (iso == NULL) {
-                    Py_DECREF(item);
-                    NpyArr_freeLabels(ret, num);
-                    ret = 0;
-                    break;
-                }
-
-                cLabel = (char *)PyUnicode_AsUTF8(iso);
-                Py_DECREF(iso);
-                len = strlen(cLabel);
-            } else {
-                npy_int64 value;
-                // TODO: refactor to not duplicate what goes on in
-                // beginTypeContext
-                if (PyObject_HasAttrString(ts, "value")) {
-                    PRINTMARK();
-                    value = get_long_attr(ts, "value");
-                } else {
-                    PRINTMARK();
-                    value = total_seconds(ts) *
-                            1000000000LL; // nanoseconds per second
-                }
-                Py_DECREF(ts);
-
-                NPY_DATETIMEUNIT unit = enc->datetimeUnit;
-                if (scaleNanosecToUnit(&value, unit) != 0) {
-                    Py_DECREF(item);
-                    NpyArr_freeLabels(ret, num);
-                    ret = 0;
-                    break;
-                }
-
-                char buf[21] = {0}; // 21 chars for 2**63 as string
-                cLabel = buf;
-                sprintf(buf, "%" NPY_INT64_FMT, value);
-                len = strlen(cLabel);
-            }
-        } else { // Fallback to string representation
+	    sprintf(cLabel, "%" PRId64, longVal);
+	    len = strlen(cLabel);
+	  }
+	}
+	else if (PyDateTime_Check(item) || PyDate_Check(item)) {
+	      NPY_DATETIMEUNIT base = enc->datetimeUnit;
+	      if (enc->datetimeIso) {
+		cLabel = PyDateTimeToIso(item, base, &len);
+	      } else {
+		sprintf(cLabel, "%" PRId64, PyDateTimeToEpoch(item, base));
+		len = strlen(cLabel);
+	      }
+	    }
+        else { // Fallback to string representation
             PyObject *str = PyObject_Str(item);
             if (str == NULL) {
                 Py_DECREF(item);
@@ -1720,7 +1705,7 @@ void Object_beginTypeContext(JSOBJ _obj, JSONTypeContext *tc) {
 
             if (enc->datetimeIso) {
                 PRINTMARK();
-                pc->PyTypeToUTF8 = NpyDateTimeToIso;
+                pc->PyTypeToUTF8 = NpyDateTimeToIsoCallback;
                 // Currently no way to pass longVal to iso function, so use
                 // state management
                 GET_TC(tc)->longValue = longVal;
@@ -1795,7 +1780,7 @@ void Object_beginTypeContext(JSOBJ _obj, JSONTypeContext *tc) {
         PRINTMARK();
         if (enc->datetimeIso) {
             PRINTMARK();
-            pc->PyTypeToUTF8 = PyDateTimeToIso;
+            pc->PyTypeToUTF8 = PyDateTimeToIsoCallback;
             tc->type = JT_UTF8;
         } else {
             PRINTMARK();