benchmarks below 

```sh
       before           after         ratio
     [9c6771c5]       [5c0f5682]
     <master>         <json-index-dates>
-        231±60ms          188±3ms     0.82  io.json.ToJSON.time_iso_format('split', 'df_td_int_ts')
-        496±20ms         218±40ms     0.44  io.json.ToJSON.time_iso_format('index', 'df_int_float_str')
-        486±20ms          207±2ms     0.42  io.json.ToJSON.time_iso_format('columns', 'df_int_float_str')
-         499±9ms          210±1ms     0.42  io.json.ToJSON.time_iso_format('index', 'df_date_idx')
-        503±20ms          206±2ms     0.41  io.json.ToJSON.time_iso_format('columns', 'df_int_floats')
-        515±20ms          210±1ms     0.41  io.json.ToJSON.time_iso_format('index', 'df')
-        524±80ms        209±0.5ms     0.40  io.json.ToJSON.time_iso_format('index', 'df_int_floats')
-        528±10ms          206±2ms     0.39  io.json.ToJSON.time_iso_format('columns', 'df_td_int_ts')
diff --git a/asv_bench/asv.conf.json b/asv_bench/asv.conf.json
index c04bbf53a..897de2f85 100644
--- a/asv_bench/asv.conf.json
+++ b/asv_bench/asv.conf.json
-        546±80ms        209±0.7ms     0.38  io.json.ToJSON.time_iso_format('index', 'df_td_int_ts')
-        568±60ms          208±3ms     0.37  io.json.ToJSON.time_iso_format('columns', 'df_date_idx')
-       598±100ms          206±2ms     0.35  io.json.ToJSON.time_iso_format('columns', 'df')

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE INCREASED.
```

Note that this mostly improves on DTI which on 0.25.3 can't even be written as ISO format, so I didn't add a whatsnew. Timedelta is the big bottleneck remaining