@jbrockmendel 2019-12-27
CI notwithstanding, this looks like a nice cleanup+speedup