@pep8speaks 2019-12-27
Hello @WillAyd! Thanks for updating this PR. We checked the lines you've touched for [PEP 8](https://www.python.org/dev/peps/pep-0008) issues, and found:

* In the file [`asv_bench/benchmarks/io/json.py`](https://github.com/pandas-dev/pandas/blob/9e510cc08308ce822a782955b3f5ff367448148e/asv_bench/benchmarks/io/json.py):

> [Line 139:16](https://github.com/pandas-dev/pandas/blob/9e510cc08308ce822a782955b3f5ff367448148e/asv_bench/benchmarks/io/json.py#L139): [E251](https://duckduckgo.com/?q=pep8%20E251) unexpected spaces around keyword / parameter equals
> [Line 139:18](https://github.com/pandas-dev/pandas/blob/9e510cc08308ce822a782955b3f5ff367448148e/asv_bench/benchmarks/io/json.py#L139): [E251](https://duckduckgo.com/?q=pep8%20E251) unexpected spaces around keyword / parameter equals
> [Line 142:10](https://github.com/pandas-dev/pandas/blob/9e510cc08308ce822a782955b3f5ff367448148e/asv_bench/benchmarks/io/json.py#L142): [E251](https://duckduckgo.com/?q=pep8%20E251) unexpected spaces around keyword / parameter equals
> [Line 142:12](https://github.com/pandas-dev/pandas/blob/9e510cc08308ce822a782955b3f5ff367448148e/asv_bench/benchmarks/io/json.py#L142): [E251](https://duckduckgo.com/?q=pep8%20E251) unexpected spaces around keyword / parameter equals
> [Line 143:14](https://github.com/pandas-dev/pandas/blob/9e510cc08308ce822a782955b3f5ff367448148e/asv_bench/benchmarks/io/json.py#L143): [E251](https://duckduckgo.com/?q=pep8%20E251) unexpected spaces around keyword / parameter equals
> [Line 143:16](https://github.com/pandas-dev/pandas/blob/9e510cc08308ce822a782955b3f5ff367448148e/asv_bench/benchmarks/io/json.py#L143): [E251](https://duckduckgo.com/?q=pep8%20E251) unexpected spaces around keyword / parameter equals
> [Line 144:14](https://github.com/pandas-dev/pandas/blob/9e510cc08308ce822a782955b3f5ff367448148e/asv_bench/benchmarks/io/json.py#L144): [E251](https://duckduckgo.com/?q=pep8%20E251) unexpected spaces around keyword / parameter equals
> [Line 144:16](https://github.com/pandas-dev/pandas/blob/9e510cc08308ce822a782955b3f5ff367448148e/asv_bench/benchmarks/io/json.py#L144): [E251](https://duckduckgo.com/?q=pep8%20E251) unexpected spaces around keyword / parameter equals
> [Line 145:19](https://github.com/pandas-dev/pandas/blob/9e510cc08308ce822a782955b3f5ff367448148e/asv_bench/benchmarks/io/json.py#L145): [E251](https://duckduckgo.com/?q=pep8%20E251) unexpected spaces around keyword / parameter equals
> [Line 145:21](https://github.com/pandas-dev/pandas/blob/9e510cc08308ce822a782955b3f5ff367448148e/asv_bench/benchmarks/io/json.py#L145): [E251](https://duckduckgo.com/?q=pep8%20E251) unexpected spaces around keyword / parameter equals
> [Line 146:18](https://github.com/pandas-dev/pandas/blob/9e510cc08308ce822a782955b3f5ff367448148e/asv_bench/benchmarks/io/json.py#L146): [E251](https://duckduckgo.com/?q=pep8%20E251) unexpected spaces around keyword / parameter equals
> [Line 146:20](https://github.com/pandas-dev/pandas/blob/9e510cc08308ce822a782955b3f5ff367448148e/asv_bench/benchmarks/io/json.py#L146): [E251](https://duckduckgo.com/?q=pep8%20E251) unexpected spaces around keyword / parameter equals
> [Line 147:16](https://github.com/pandas-dev/pandas/blob/9e510cc08308ce822a782955b3f5ff367448148e/asv_bench/benchmarks/io/json.py#L147): [E251](https://duckduckgo.com/?q=pep8%20E251) unexpected spaces around keyword / parameter equals
> [Line 147:18](https://github.com/pandas-dev/pandas/blob/9e510cc08308ce822a782955b3f5ff367448148e/asv_bench/benchmarks/io/json.py#L147): [E251](https://duckduckgo.com/?q=pep8%20E251) unexpected spaces around keyword / parameter equals