@jreback 2020-01-02
thanks, might be worth a general whatsnew note for perf for json changes (e.g. list issues that impacted perf that you have done recently)