@WillAyd 2020-01-02
Yea no problem; will address in follow up